# Cache blockly flyouts for faster subsequent loads (~3X speedup)