--- a/localtypings/blockly.d.ts
+++ b/localtypings/blockly.d.ts
@@ -1370,6 +1370,25 @@ declare module Blockly {
             setCommentText(text: string): void;
     
             /**
+             * Set this Block's breakpoint as enabled or disabled.
+             * @param {boolean} enable A boolean definining if the breakpoint should be enabled or disabled.
+             */
+            enableBreakpoint(enable: boolean): void;
+    
+            /**
+             * Returns a boolean representing if the block has a breakpoint set or not 
+             * (regardless of whether it is enabled).
+             * @return {boolean} Block's breakpoint set or not.
+             */
+            isBreakpointSet(): boolean;
+    
+            /**
+             * Set this block's breakpoint.
+             * @param {boolean} set Boolean representing if the breakpoint is now set or not.
+             */
+            setBreakpoint(set: boolean): void;
+    
+            /**
              * Set this block's output shape.
              * e.g., null, OUTPUT_SHAPE_HEXAGONAL, OUTPUT_SHAPE_ROUND, OUTPUT_SHAPE_SQUARE.
              * @param {?number} outputShape Value representing output shape
@@ -1670,6 +1689,13 @@ declare module Blockly {
              *     being moved to a different surface.
              */
             clearAndHide(opt_newSurface?: Element): void;
+    
+            /**
+             * Sets the opacity of the drag surface and everything on it.
+             * @param {number} value The new opacity value to use.
+             * @package
+             */
+            setOpacity(value: number): void;
     } 
     
 }
@@ -2727,7 +2753,13 @@ declare module Blockly {
             warning: Blockly.Warning;
     
             /**
-             * Returns a list of mutator, comment, and warning icons.
+             * Block's breakpoint icon (if any).
+             * @type {Blockly.Breakpoint}
+             */
+            breakpoint: Blockly.Breakpoint;
+    
+            /**
+             * Returns a list of breakpoint, mutator, comment, and warning icons.
              * @return {!Array} List of icons.
              */
             getIcons(): any[];
@@ -2938,6 +2970,12 @@ declare module Blockly {
              * @return {string} Block's comment.
              */
             getCommentText(): string;
+    
+            /**
+             * Set this blockSvg's breakpoint as enabled.
+             * @param {boolean} enable Boolean representing if the breakpoint is being enabled or disabled.
+             */
+            enableBreakpoint(enable: boolean): void;
     
             /**
              * Set this block's comment text.
@@ -5155,9 +5193,9 @@ declare module Blockly.DropDownDiv {
 
     /**
      * Provide the div for inserting content into the drop-down.
-     * @return {Element} Div to populate with content
+     * @return {HTMLElement} Div to populate with content
      */
-    function getContentDiv(): Element;
+    function getContentDiv(): HTMLElement;
 
     /**
      * Clear the content of the drop-down.
@@ -5797,10 +5835,10 @@ declare module Blockly {
     
             /**
              * Box drawn around a field.
-             * @type {SVGRectElement}
+             * @type {SVGElement}
              * @private
              */
-            arrow_: SVGRectElement;
+            arrow_: SVGElement;
     
             /**
              * Box drawn around a field.
@@ -10102,6 +10140,14 @@ declare module Blockly {
             renderIcon(cursorX: number): number;
     
             /**
+             * Move the icon.
+             * @param {number} cursorX Horizontal offset at which to position the icon.
+             * @param {number} cursorY Vertical offset at which to position the icon.
+             * @return {number} Horizontal offset for next item to draw.
+             */
+            moveIcon(cursorX: number, cursorY: number): number;
+    
+            /**
              * Notification that the icon has moved.  Update the arrow accordingly.
              * @param {!goog.math.Coordinate} xy Absolute location in workspace coordinates.
              */
@@ -13441,9 +13487,10 @@ declare module Blockly.Functions {
      * Create a flyout, creates the DOM elements for the flyout, and initializes the flyout.
      * @param {!Blockly.Workspace} workspace The target and parent workspace for this flyout. The workspace's options will
      *     be used to create the flyout's inner workspace.
-     * @return { flyout: !Blockly.Flyout, flyoutSvg: SVGSVGElement } True if the block is a function argument reporter.
+     * @param {!Element} siblingNode The flyout is added after this reference node. 
+     * @return {!Blockly.Flyout} The newly created flyout.
      */
-    function createFlyout(workspace: Blockly.Workspace): void;
+    function createFlyout(workspace: Blockly.Workspace, siblingNode: Element): Blockly.Flyout;
 }
 
 declare module Blockly.pxtBlocklyUtils {
@@ -13480,6 +13527,69 @@ declare module Blockly.pxtBlocklyUtils {
      * @return {boolean} Whether there is a matching reporter or not.
      */
     function hasMatchingArgumentReporter(targetBlock: Blockly.Block, reporter: Blockly.Block): boolean;
+}
+
+declare module Blockly {
+
+    class Breakpoint extends Breakpoint__Class { }
+    /** Fake class which should be extended to avoid inheriting static properties */
+    class Breakpoint__Class extends Blockly.Icon__Class  { 
+    
+            /**
+             * Class for a breakpoint.
+             * @param {!Blockly.Block} block The block associated with this breakpoint.
+             * @extends {Blockly.Icon}
+             * @constructor
+             */
+            constructor(block: Blockly.Block);
+    
+            /**
+             * Does this icon get hidden when the block is collapsed.
+             */
+            collapseHidden: any /*missing*/;
+    
+            /**
+             * Create the icon on the block.
+             */
+            createIcon(): void;
+    
+            /**
+             * Draw the breakpoint icon.
+             * @param {!Element} group The icon group.
+             * @private
+             */
+            drawIcon_(group: Element): void;
+    
+            /**
+             * Enable/Disable the breakpoint icon.
+             * @private
+             */
+            enableBreakpoint(): void;
+    
+            /**
+             * Dispose of this breakpoint Icon.
+             */
+            dispose(): void;
+    
+            /**
+             * Toggle the breakpoint icon between set and unset.
+             * @param {boolean} visible True if the breakpoint icon should be set.
+             */
+            setVisible(visible: boolean): void;
+    
+            /**
+             * Is this breakpoint set?
+             * @return {boolean} True if the breakpoint is Set.
+             */
+            isVisible(): boolean;
+    
+            /**
+             * Notification that the icon has moved.
+             * @param {!goog.math.Coordinate} xy Absolute location in workspace coordinates.
+             */
+            setIconLocation(xy: goog.math.Coordinate): void;
+    } 
+    
 }
 
 declare module Blockly.PXTUtils {
@@ -14880,7 +14990,7 @@ declare module Blockly.Events {
             /**
              * Class for a UI event.
              * @param {Blockly.Block} block The affected block.
-             * @param {string} element One of 'selected', 'comment', 'mutator', etc.
+             * @param {string} element One of 'selected', 'comment', 'mutator', 'breakpoint' etc.
              * @param {*} oldValue Previous value of element.
              * @param {*} newValue New value of element.
              * @extends {Blockly.Events.Abstract}
@@ -18106,9 +18216,9 @@ declare module Blockly {
     
             /**
              * Get the SVG element that contains this workspace.
-             * @return {Element} SVG element.
+             * @return {SVGElement} SVG element.
              */
-            getParentSvg(): Element;
+            getParentSvg(): SVGElement;
     
             /**
              * Translate this workspace to new coordinates.