… webapp/public/blockly/blockly_compressed.js