REQUIRES: https://github.com/Microsoft/pxt-blockly/pull/208

Previously, we would recompute and relayout all the blocks each time we opened the "flyout" when clicking on the toolbox.

With this change, we maintain separate flyouts for each toolbox category and toggle their visiblity state when switching between categories. This doesn't speed up the initial load, but subsequent loads are ~3X faster.

I'm also investigating a way to get another 3X (for a total of 9X) speed up by playing nice with React so it doesn't force a style recompute when toggling visibility.

Note this affects all blockly editors.

## Before
[macbookpro, x6 slowdown] 1st load: ~1.3 seconds
[macbookpro, x6 slowdown] nth load: ~600 ms
[chrome book LC330] 1st load: ~1.8 seconds
[chrome book LC330] nth load: ~680ms
![flyouts_perf_before](https://user-images.githubusercontent.com/6453828/54047374-2ec5cd80-41d7-11e9-9bc9-7c15c4b0bd94.gif)

## After
[macbookpro, x6 slowdown] 1st load: ~1.3 seconds (same)
[macbookpro, x6 slowdown] nth load: ~180 ms
![flyouts_perf_after](https://user-images.githubusercontent.com/6453828/54047379-35544500-41d7-11e9-8136-88918f80c466.gif)

perf numbers are from Chrome profiler using: 
- 6x CPU throttling on MacBook Pro 2018 i9
- Lenovo C330 Chrome book w/ MT8173C