@pelikhan 2019-03-08
Do you clear the cache when the header gets reloaded?