@darzu 2019-03-11
> Do you clear the cache when the header gets reloaded?

What do you mean by "header"?