@pelikhan 2019-03-11
When a script gets reloaded. Looks like you do.

See “loadHeaderAsync”

Get Outlook for iOS<https://aka.ms/o0ukef>
________________________________
From: Daryl Zuniga <notifications@github.com>
Sent: Monday, March 11, 2019 12:40:07 AM
To: Microsoft/pxt
Cc: Peli de Halleux; Comment
Subject: Re: [Microsoft/pxt] Cache blockly flyouts for faster subsequent loads (~3X speedup) (#5323)


Do you clear the cache when the header gets reloaded?

What do you mean by "header"?

—
You are receiving this because you commented.
Reply to this email directly, view it on GitHub<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FMicrosoft%2Fpxt%2Fpull%2F5323%23issuecomment-471434367&data=02%7C01%7Cjhalleux%40microsoft.com%7C49bc4011f4e64f42dc8f08d6a5f4c928%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636878868147613692&sdata=LY5IUs%2FmlAMjL6FSL9qNP2GFVC4Pay%2FsyHxNaSa1gaM%3D&reserved=0>, or mute the thread<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FAD-4KVmKtzLPxLzCqvKmz8PB3OHqLsJzks5vVghXgaJpZM4bl4G3&data=02%7C01%7Cjhalleux%40microsoft.com%7C49bc4011f4e64f42dc8f08d6a5f4c928%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636878868147613692&sdata=FNHSP3hGparZQBE0Ebox0qn8OTseMD8YZLeVSxL%2BDxI%3D&reserved=0>.