@darzu 2019-03-12
Flyout perf improvements testable here: https://arcade.staging.pxt.io/app/ccc806ea325942a07d3e7b76c24dfbbc41876c9a-e6714039f6