@darzu 2019-03-12
> Ready to go?

Not yet. The staging deployment I made of these changes is failing with:
> {"message":"Internal server error (catch)"}
and I want to make sure it's not related to my changes.

I'll merge when rdy