@darzu 2019-03-12
Testable here: https://arcade.makecode.com/app/c641ae5a704f37f9bfe91a1f4f388e4fb56b94e2-df2a70577f#editor